remove the decimal point in product prices

+++ POS_Shop/Models/Product.cs
@@ -17,8 +17,8 @@ namespace POS_Shop.Models
         public string ProductUrduName { get; set; }
 
         public string ProductType { get; set; }
-        public decimal? PurchasePrice { get; set; }
-        public decimal? SalePrice { get; set; }
+        public int? PurchasePrice { get; set; }
+        public int? SalePrice { get; set; }
 
         public int? Cost { get; set; }
